Final Solution To the BDDTask.
