+++ LocatorsTests/UITests/GlobalSearchTests.cs
@@ -1,7 +1,7 @@
 using Tests.Base;
 using Business.Pages;
 
-namespace Tests.TAF
+namespace Tests.UITests
 {
     [TestFixture]
     public class GlobalSearchTests : TestBase
@@ -19,7 +19,7 @@ namespace Tests.TAF
                 .Search(term)
                 .ClickFindButton();
 
-            Assert.That(globalSearchPage.AllResultsContain(term), $"All results should contain: {term}", true);
+            Assert.That(globalSearchPage.AllResultsContain(term), $"All results should contain: {term}");
         }
     }
 }
